pyproject.toml
[tool.poetry]
name = "hw-sql"
version = "0.1.0"
description = ""
authors = ["Aleksandr <aleksandr.voitushenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
faker = "^26.0.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
